refactor : 무한 스크롤 코드 분리
